+++ substra/sdk/backends/local/backend.py
@@ -852,6 +852,8 @@ class Local(base.BaseBackend):
 def _output_from_spec(outputs: List[schemas.ComputeTaskOutput]) -> List[models.ComputeTaskOutput]:
     """Convert a list of schemas.ComputeTaskOuput to a list of models.ComputeTaskOutput"""
     return {
-        identifier: models.ComputeTaskOutput(permissions=models.Permissions(process=output.permissions))
+        identifier: models.ComputeTaskOutput(
+            permissions=models.Permissions(process=output.permissions), value=None
+        )  # default isNone (= outputs are not computed yet)
         for identifier, output in outputs.items()
     }